shape ops work in progress
git-svn-id: http://skia.googlecode.com/svn/trunk@4029 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ experimental/Intersection/LineParameterization.cpp
@@ -37,7 +37,7 @@ bool implicit_matches_ulps(const _Line& one, const _Line& two, int ulps) {
           (dy1 * dy2) * dx1 / dy1 == (dy1 * dy2) * dx2 / dy2
                  dy2  * dx1       ==  dy1        * dx2
      */
-    int diff = UlpsDiff(oneD.x * twoD.y, twoD.x * oneD.y);
+    int diff = UlpsDiff((float) (oneD.x * twoD.y), (float) (twoD.x * oneD.y));
     if (diff < 0 || diff > ulps) {
         return false;
     }
@@ -46,8 +46,8 @@ bool implicit_matches_ulps(const _Line& one, const _Line& two, int ulps) {
          dx * (y0 - x0 * dy / dx) == dx * (y1 - x1 * dy / dx)
          dx *  y0 - x0 * dy       == dx *  y1 - x1 * dy
      */
-    diff = UlpsDiff(oneD.x * one[0].y - oneD.y * one[0].x,
-            oneD.x * two[0].y - oneD.y * two[0].x);
+    diff = UlpsDiff((float) (oneD.x * one[0].y - oneD.y * one[0].x),
+            (float) (oneD.x * two[0].y - oneD.y * two[0].x));
     return diff >= 0 && diff <= ulps;
 }
